Fix handling ints for non-nested hash table option
`.map` is called on value of options passed in from `create_table` so\
that complex nested hash options can be parsed into proper Cassandra
syntax. However, this caused integer values on simple non-nested hash
options to break as `.map` was being called on these integer values. See
PR #77.

Change logic to check that value is a hash (indicating that a nested
hash was passed into options) before parsing it into a hash with string
keys and values. If it isn't then use the value as is.

Add tests for PR #77.