Update cassandra_migrations.gemspec
For now the best place to put your name and contact is on the README.md file as a contributor (also to be fair with other contributors). As I said earlier, if you wish to help maintain the gem on the long term I see no problem in putting you as an author too.